frontend/api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-cloud-firestore==2.12.0
google-cloud-monitoring==2.15.1
google-auth==2.23.0
gunicorn==21.2.0
python-dateutil==2.8.2